+++ src/tileentity/OreDeposit.cpp
@@ -111,8 +111,8 @@ namespace Game3 {
 			sprite_renderer(*texture, {
 				.x = static_cast<float>(position.column),
 				.y = static_cast<float>(position.row),
-				.x_offset = x / 2.f,
-				.y_offset = y / 2.f,
+				.x_offset = static_cast<float>(x) / 2.f,
+				.y_offset = static_cast<float>(y) / 2.f,
 				.size_x = static_cast<float>(tilesize),
 				.size_y = static_cast<float>(tilesize),
 			});
